+++ test/core/test_periodic_verlet_list.cpp
@@ -138,14 +138,14 @@ BOOST_AUTO_TEST_CASE(test_VerletList_PeriodicBoundary)
             {
                 // should be enough distant (>= threshold)
                 const auto dist = mjolnir::math::length(sys.adjust_direction(
-                            sys.position(j) - sys.position(i)));
+                            sys.position(i), sys.position(j)));
                 BOOST_TEST(dist >= threshold);
             }
             else
             {
                 // should be enough close (< threshold)
                 const auto dist = mjolnir::math::length(sys.adjust_direction(
-                            sys.position(j) - sys.position(i)));
+                            sys.position(i), sys.position(j)));
                 BOOST_TEST(dist < threshold);
             }
         }
@@ -236,7 +236,7 @@ BOOST_AUTO_TEST_CASE(test_VerletList_PeriodicBoundary_partial)
             {
                 // should be enough distant (>= threshold)
                 const auto dist = mjolnir::math::length(sys.adjust_direction(
-                            sys.position(j) - sys.position(i)));
+                            sys.position(i), sys.position(j)));
                 const bool enough_distant = dist >= threshold;
 
                 // or not a participant
@@ -255,7 +255,7 @@ BOOST_AUTO_TEST_CASE(test_VerletList_PeriodicBoundary_partial)
 
                 // should be enough close (< threshold)
                 const auto dist = mjolnir::math::length(sys.adjust_direction(
-                            sys.position(j) - sys.position(i)));
+                            sys.position(i), sys.position(j)));
                 BOOST_TEST(dist < threshold);
             }
         }
@@ -337,7 +337,7 @@ BOOST_AUTO_TEST_CASE(test_VerletList_PeriodicBoundary_partial_2)
             {
                 // should be enough distant (>= threshold)
                 const auto dist = mjolnir::math::length(sys.adjust_direction(
-                            sys.position(j) - sys.position(i)));
+                            sys.position(i), sys.position(j)));
                 const bool enough_distant = dist >= threshold;
 
                 // or not a participant
@@ -356,7 +356,7 @@ BOOST_AUTO_TEST_CASE(test_VerletList_PeriodicBoundary_partial_2)
 
                 // should be enough close (< threshold)
                 const auto dist = mjolnir::math::length(sys.adjust_direction(
-                            sys.position(j) - sys.position(i)));
+                            sys.position(i), sys.position(j)));
                 BOOST_TEST(dist < threshold);
             }
         }