+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Change Log
 
+## 1.19.0
+
+### Added
+- support OpenMP + g-BAOAB
+- output version and compiler information in log
+
+### Changed
+- make boundary condition more customizable
+
+## 1.18.0
+
+### Added
+
+- allow list of indices in offset (@yutakasi634)
+- enable to load random number generator state from checkpoint file
+
+### Changed
+
+- update toml11 to v3.6.0
+
+### Fixed
+
+- avoid error in an edge case in PeriodicCellList
+- stabilize FLP angle potential calculation in an edge case
+
+## 1.17.0
+
+### Changed
+
+- update toml11 to v3.5.0
+- speedup EnergyCalculation simulator
+
+### Fixed
+
+- read boundary information correctly in EnergyCalculation simulator
+
+### Misc
+
+- refactor energy format methods
+
 ## 1.16.0
 
 ### Added